Strange Camera and Player Implemented
use the StrangePlayer prefab as a starting point: it contains some basic gameobject hierarchy to help with scene management later and also contains a script to move the player around with the WASD keys.

use the StrangeCamera prefab to look around the scene. set the target to the "Main Camera Target" on the player, and also set the player to the strange player prefab you just dragged onto the scene