Interaction Engine added
to start: drag the Interaction Engine on to the scene and then drag in the appropriate objects to the script (read the headers)

Chat system more refined: infinite loops fixed and can work with the interaction system

a template class has been created for interactibles to complete most of the boiler plat code for you.

new interaction engine handles both collider detection and UI management of the Interaction icon.