players can no linger run a dialogue choice twice unless allowMultipleChoices is set to true
fixed #1
fixed #2

when the user runs a dialogue choice for the first time, their choice is remembered.

the second time they run that dialogue, the choice wheel is not shown and instead the tooltip is shown and when they press the chat key, the choice they picked last time is run.

every dialogue has a firstRun bool. once it is run, this is set to false.
when false. when a dialogue runs, ginving and turning in quests are disabled

stops duplicate quest giving, and is a better solution than the last commit

if you wish to disable this and allow a player to choose/run a dialogue choice twice. just change allowMultipleChoices to true