beginning to add the option to implement custom behaviour + flying enemies coming soon
most functions have been made protected and virtual so that the user can override my controls to add additional behaviour or overwrite my behaviour with their own.

while reformatting the code to make it more suitable for a user to override, it randomly occurred to my that they may want their enemies to fly (or just move in the y axis) and not be glued to the ground.

I started playing around with the movement code and made a 3d version. I think I am going to dedicate some time to making flying enemy AI since I find it interesting