Map Gen: instead of calculating colour per-face. we now calculate it per vertex
each vertex will calculate its gradient by getting the y values of itself and 8 8 of its neighbours (wherever possible). then the range  is found which is a symbolic value that represents the gradient of a vertex (how steep is the hill)

then the colours are assigned linearly  using an inverse Lerp. however sometimes this  lerp has anomalies that make most of the land the same colour. a Sensitivity bar is used to push the values further up the spectrum and changes how much the steepness of a hill affects the colour of its vertices