Map gen can now generate multiple prefabs
map gen now instantiates a prefab and generates the mesh on that prefab. this means that multiple meshes can now be spawned in programattically.

however. there is an issue with stitching.  because each chunk is only aware of its own values, when normalised, gaps appear around the edges of chunks because each chunk is using a different max and min to normalise