Map gen no longer has gaps between chunks
unfortunately my method of applying the home chunk's min and max's to the InverseLerp (for normalization) didnt work as numbers > 1 were capped at 1.

instead i decided to go with calculating the theoretical min and maxes. this worked a lot better. however, even with consistent normalisation across the whole map. there was still a problem with small gaps.

this was because the spacing between chunks was incorrect and as 1 map unit too far... a simple -1 fixed this :)

fixes #18