colour stitching and detail zooming
colour stitching now uses a global min and max gradient.

theoretical min and max set to 0 and size*2.
closes #24

mapgen now uses a "detail" option which effectively zoom in on the noise.
if you have a map you like, but you can still see the triangles, and it looks a little lowpoly, you can zoom in on the noise to have the exact same map but zoomed in. you may need to add more chunks

closes #22