noise now has a layer of padding
this layer of padding is never rendered, but acts as an overlap for chunks.

this way the edges of chunks can have somewhat constant colouring between seams.

I have found this works best on large chunks. smaller chunks still have some sort of issue but I think  this is a problem with the UVs/lighting instead of the colouring